requirements.txt
libclang>=16.0.0
requests>=2.28.0
